+++ src/main/java/duke/command/ByeCommand.java
@@ -1,3 +1,10 @@
+package duke.command;
+
+import duke.Storage;
+import duke.TaskList;
+import duke.Ui;
+import duke.exception.DukeCommandException;
+
 public class ByeCommand extends Command{
     public ByeCommand(String inputCommand) {
         super(inputCommand);
