+++ src/main/java/duke/exception/DukeStorageException.java
@@ -1,6 +1,8 @@
+package duke.exception;
+
 public class DukeStorageException extends DukeException{
 
-    DukeStorageException(String message) {
+    public DukeStorageException(String message) {
         super(message);
     }
 }
