+++ src/main/java/duke/util/SerializeUtil.java
@@ -1,3 +1,5 @@
+package duke.util;
+
 import java.io.*;
 
 public class SerializeUtil {
