+++ text-ui-test/EXPECTED.TXT
@@ -4,28 +4,38 @@
         ____________________________________________________________
 
         ____________________________________________________________
+        Here are the tasks in your list:
+        ____________________________________________________________
+
+        ____________________________________________________________
         Got it. I've added this task:
-          [T][✘] read book
+          [T][✘] aaa
         Now you have 1 task(s) in the list.
         ____________________________________________________________
 
         ____________________________________________________________
-        Got it. I've added this task:
-          [D][✘] return book (by: tonight)
-        Now you have 2 task(s) in the list.
+        ☹ OOPS!!! Wrong 'deadline' command format!
         ____________________________________________________________
 
         ____________________________________________________________
-        Got it. I've added this task:
-          [E][✘] play games (at: tomorrow 9-11pm)
-        Now you have 3 task(s) in the list.
+        ☹ OOPS!!! Wrong 'deadline' command format!
+        ____________________________________________________________
+
+        ____________________________________________________________
+        ☹ OOPS!!! Wrong 'event' command format!
+        ____________________________________________________________
+
+        ____________________________________________________________
+        ☹ OOPS!!! Wrong 'event' command format!
+        ____________________________________________________________
+
+        ____________________________________________________________
+        ☹ OOPS!!! Wrong 'event' command format!
         ____________________________________________________________
 
         ____________________________________________________________
         Here are the tasks in your list:
-        1.[T][✘] read book
-        2.[D][✘] return book (by: tonight)
-        3.[E][✘] play games (at: tomorrow 9-11pm)
+        1.[T][✘] aaa
         ____________________________________________________________
 
         ____________________________________________________________
@@ -33,56 +43,18 @@
         ____________________________________________________________
 
         ____________________________________________________________
-        ☹ OOPS!!! Wrong 'todo' command format!
-        ____________________________________________________________
+        Nice! I've marked this task as done:
 
-        ____________________________________________________________
-        ☹ OOPS!!! Wrong 'todo' command format!
-        ____________________________________________________________
-
-        ____________________________________________________________
-        ☹ OOPS!!! Wrong 'deadline' command format!
-        ____________________________________________________________
-
-        ____________________________________________________________
-        ☹ OOPS!!! Wrong 'deadline' command format!
-        ____________________________________________________________
-
-        ____________________________________________________________
-        ☹ OOPS!!! Wrong 'event' command format!
-        ____________________________________________________________
-
-        ____________________________________________________________
-        ☹ OOPS!!! Wrong 'event' command format!
-        ____________________________________________________________
-
-        ____________________________________________________________
-        Here are the tasks in your list:
-        1.[T][✘] read book
-        2.[D][✘] return book (by: tonight)
-        3.[E][✘] play games (at: tomorrow 9-11pm)
-        ____________________________________________________________
-
-        ____________________________________________________________
-        Got it. I've added this task:
-          [T][✘] illegal stuff Kappa
-        Now you have 4 task(s) in the list.
-        ____________________________________________________________
-
-        ____________________________________________________________
-        ☹ OOPS!!! There isn't a task with that index!
+          [T][✓] aaa
         ____________________________________________________________
 
         ____________________________________________________________
         Noted. I've removed this task:
-          [T][✘] illegal stuff Kappa
+          [T][✓] aaa
         ____________________________________________________________
 
         ____________________________________________________________
         Here are the tasks in your list:
-        1.[T][✘] read book
-        2.[D][✘] return book (by: tonight)
-        3.[E][✘] play games (at: tomorrow 9-11pm)
         ____________________________________________________________
 
         ____________________________________________________________
